[Button] adjust center of circle button as its fixed for software vector backend.

+++ src/lib/tizen_vector.c
@@ -1158,8 +1158,8 @@ button_effect_no_bg_resize_cb(void *data, Evas *e EINA_UNUSED,
    evas_vg_shape_shape_reset(vd->shape[0]);
    if (vd->is_circle)
      {
-        int radius_w = w / 2;
-        int radius_h = h / 2;
+        double radius_w = w / 2;
+        double radius_h = h / 2;
         evas_vg_shape_shape_append_circle(vd->shape[0], radius_w, radius_h, radius_w);
      }
    else
@@ -1262,8 +1262,8 @@ button_effect_resize_cb(void *data, Evas *e EINA_UNUSED,
    evas_vg_shape_shape_reset(vd->shape[1]);
    if (vd->is_circle)
      {
-        int radius_w = w / 2;
-        int radius_h = h / 2;
+        double radius_w = w / 2;
+        double radius_h = h / 2;
         evas_vg_shape_shape_append_circle(vd->shape[1], radius_w, radius_h, radius_w);
      }
    else
@@ -1293,8 +1293,8 @@ button_base_resize_cb(void *data, Evas *e EINA_UNUSED,
    evas_vg_shape_shape_reset(vd->shape[0]);
    if (vd->is_circle)
      {
-        int radius_w = w / 2;
-        int radius_h = h / 2;
+        double radius_w = w / 2;
+        double radius_h = h / 2;
         evas_vg_shape_shape_append_circle(vd->shape[0], radius_w, radius_h, radius_w);
      }
    else
